+++ AddOns/WclPlayerScore-WotLK/CHANGELOG.md
@@ -1,10 +1,10 @@
 # WclPlayerScore-WotLK
 
-## [v1.2.10](https://github.com/icaca/WclPlayerScore/tree/v1.2.10) (2022-11-08)
-[Full Changelog](https://github.com/icaca/WclPlayerScore/commits/v1.2.10) [Previous Releases](https://github.com/icaca/WclPlayerScore/releases)
+## [v1.2.12](https://github.com/icaca/WclPlayerScore/tree/v1.2.12) (2022-11-10)
+[Full Changelog](https://github.com/icaca/WclPlayerScore/commits/v1.2.12) [Previous Releases](https://github.com/icaca/WclPlayerScore/releases)
 
 - WCL Data Update  
-- Update README.md  
 - WCL Data Update  
 - WCL Data Update  
 - Update README.md  
+- WCL Data Update